Bugfix FXIOS-32351 Avoid unnecessary URLSession to obtain certificates
